src/app/api/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
bcrypt==4.2.1
beanie==1.27.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
flake8==7.1.1
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.16.0
http_ece==1.2.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
lazy-model==0.2.0
mccabe==0.7.0
mercadopago==2.2.3
motor==3.6.0
multidict==6.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
py-vapid==1.9.2
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pyflakes==3.2.0
PyJWT==2.10.0
pymongo==4.9.2
pyotp==2.9.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pywebpush==2.0.3
qrcode==8.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
toml==0.10.2
typing_extensions==4.12.2
Unidecode==1.4.0
urllib3==2.2.3
uvicorn==0.32.1
websockets==14.1
yarl==1.18.3
